Bump Ansible to 2.9.6
Support for newer Ansible versions have been introduced for Bifrost
and Kubespray recently so this change uplifts Ansible to 2.9.6.
However, Ansible 2.9.6 dependencies require newer versions of some of the
packages for offline deployments.

This results in moving to more recent versions of various python packages
causing conflicts for Bifrost dependency constraints. In the end, everything
required to be updated in lockstep.

Here is the list
- bifrost, its dependencies, and requirements are bumped due to the reasons
listed above.
- pip is uplifted to 20.0.2 due to pep517.
- no-use-pep517 is introduced as workaround for pip install since setuptools
is not recognized as build dependency even though it is available.
- kubespray is bumped due to change in filters in new Ansible version.
- conntrack and python3-mysqldb are added as new packages to include in offline
package due to upstream.
- vbmc 2.0 dropped the daemon autostart so we need to start it explicitly
via systemd by creating systemd unit config file.

Another thing to note is that there is a temporary workaround to overwrite
bifrost bootstrap.yml which will be removed once upstream is made aware of
the issue and a fix is implemented.

Finally, the ability to override Bifrost and Ironic versions via environment
variables is removed as well since it is not used.

Change-Id: I0ff8a84bc7238e8a6405c55bee65bcfedd0fae57