Package the patch for testing it in CI
In order to ensure a change that is sent to an engine
repo is tested properly for packaging and offline
deployments, the change/patch must be included in the
package instead of tip of the master of the corresponding
repository. This change updates build script to achieve
this so if GERRIT_CHANGE_ID is set, build script determines
the Gerrit project the change/patch is sent for, navigates
the clone of the repo in /tmp/offline-package and checks out
the patch to package it.

Change-Id: I2fadb460fffc298f88b297ba691ab1f2c605f8c2