Rename packaging playbook and role
As we are starting to work on enabling packaging for more
stacks than just Kubernetes, it would make sense to have common
approach to different stuff we do.

For installation, the apps have a playbook and corresponding role
named install. For packaging, we can follow the same approach with
a playbook and corresponding role named package which is what this
change does by renaming kubespray specific packaging playbook and
the role.

Change-Id: I71f19f0247771fcdef53b18030348aaf50fcef11